% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/can.R
\name{can}
\alias{can}
\title{Percentage of cannibals}
\usage{
can(graph)
}
\arguments{
\item{graph}{an igraph object.}
}
\value{
numeric: Percentage of cannibals
}
\description{
Percentage of cannibals species in a food web.
}
\examples{
data(aleutian)

can(aleutian)
}
\references{
Baiser, B., Gotelli, N. J., Buckley, H. L., Miller, T. E., & Ellison, A. M. (2012). Geographic variation in network structure of a nearctic aquatic food web. Global Ecology and Biogeography.
}
