% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality.R
\name{closeness_centrality}
\alias{closeness_centrality}
\title{Closeness centrality}
\usage{
closeness_centrality(graph, ...)
}
\arguments{
\item{graph}{an \code{igraph} object.}

\item{...}{others arguments passed to \code{\link[=closeness]{igraph::closeness()}}}
}
\value{
\code{numeric} a vector containing closeness centrality coefficients for each node.
}
\description{
Closeness centrality is a measure of the proximity of a node to all others.
}
\references{
Delmas, E., Besson, M., Brice, M.-H., Burkle, L. A., Dalla Riva, G. V., Fortin, M.-J., … Poisot, T. (2019). Analysing ecological networks of species interactions. Biological Reviews.
}
