% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectance.R
\name{connectance}
\alias{connectance}
\title{Connectance of a graph}
\usage{
connectance(graph, loops = TRUE, directed = TRUE)
}
\arguments{
\item{graph}{an igraph object.}

\item{loops}{logical: are pecies able to interact with themselves?}

\item{directed}{logical: are interactions directed?}

\item{unipartite}{not implemented yet}
}
\value{
numeric: the value of the connectance.
}
\description{
Connectance of a graph
}
