% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality.R
\name{degree_centrality}
\alias{degree_centrality}
\title{Degreee centrality}
\usage{
degree_centrality(graph, ...)
}
\arguments{
\item{graph}{an \code{igraph} object.}

\item{...}{others arguments passed to \code{\link[=degree]{igraph::degree()}}}
}
\value{
\code{numeric} a vector containing the degree of each node.
}
\description{
Degree centrality is the count of the number of interaction for each node. It can be partitioned between in-degree and out-degree for directed networks.
}
\examples{
data(aleutian)

degree_centrality(aleutian)
}
\references{
Delmas, E., Besson, M., Brice, M.-H., Burkle, L. A., Dalla Riva, G. V., Fortin, M.-J., … Poisot, T. (2019). Analysing ecological networks of species interactions. Biological Reviews.
}
