% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degree_distribution.R
\name{degree_distribution}
\alias{degree_distribution}
\title{Degree distribution}
\usage{
degree_distribution(graph, ...)
}
\arguments{
\item{graph}{an igraph object.}

\item{...}{others arguments passed to \code{\link[=degree]{igraph::degree_distribution}}}
}
\value{
named numeric : a numeric vector of the degree distribution.
}
\description{
Compute the degree distribution of a network
}
\examples{
data(aleutian)

degree_distribution(aleutian)
}
\references{
Delmas, E., Besson, M., Brice, M.-H., Burkle, L. A., Dalla Riva, G. V., Fortin, M.-J., … Poisot, T. (2019). Analysing ecological networks of species interactions. Biological Reviews.
}
