% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diameter.R
\name{diameter}
\alias{diameter}
\title{Diameter of a network}
\usage{
diameter(graph, ...)
}
\arguments{
\item{graph}{an \code{igraph} object.}

\item{...}{others arguments passed to \code{\link[igraph::diameter]{igraph::degree.distribution}}}
}
\value{
\code{numeric}: the diameter of the graph.
}
\description{
Diameter of a network.
}
\references{
Delmas, E., Besson, M., Brice, M.-H., Burkle, L. A., Dalla Riva, G. V., Fortin, M.-J., … Poisot, T. (2019). Analysing ecological networks of species interactions. Biological Reviews.
}
