% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality.R
\name{eigenvector_centrality}
\alias{eigenvector_centrality}
\title{Eigenvector centrality}
\usage{
eigenvector_centrality(graph, full = FALSE, ...)
}
\arguments{
\item{graph}{an \code{igraph} object.}

\item{full}{\code{logical} should all the \code{\link[=eigen_centrality]{igraph::eigen_centrality()}} output be returned or only the scores?}

\item{...}{others arguments passed to \code{\link[=eigen_centrality]{igraph::eigen_centrality()}}}
}
\value{
\code{numeric} a vector containing eigenvector centrality coefficients for each node.
\code{list} see \code{\link[=eigen_centrality]{igraph::eigen_centrality()}} documentation for more informations.
}
\description{
Eigenvector centrality is a meaure of the relative importance of each node in the network by giving a score based on the nomber of links of each node.
}
\examples{
data(aleutian)

eigenvector_centrality(aleutian)
}
\references{
Delmas, E., Besson, M., Brice, M.-H., Burkle, L. A., Dalla Riva, G. V., Fortin, M.-J., … Poisot, T. (2019). Analysing ecological networks of species interactions. Biological Reviews.
}
