% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy.R
\name{gen}
\alias{gen}
\title{Generality}
\usage{
gen(graph)
}
\arguments{
\item{graph}{an igraph object.}
}
\value{
dataframe: name of the taxon and its number of prey
}
\description{
The number of prey of a taxon.
}
\examples{
data(aleutian)

gen(aleutian)
}
\references{
Williams, R. J., & Martinez, N. D. (2000). Simple rules yield complex food webs. Nature.
}
