% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy.R
\name{gen_sd}
\alias{gen_sd}
\title{Normalized standard deviation of generality}
\usage{
gen_sd(graph)
}
\arguments{
\item{graph}{an igraph object.}
}
\value{
numeric: the normalized standard deviation of generality.
}
\description{
The number of prey of a taxon standardized by L/S.
}
\examples{
data(aleutian)

gen_sd(aleutian)
}
\references{
Williams, R. J., & Martinez, N. D. (2000). Simple rules yield complex food webs. Nature.
}
