% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herbivory.R
\name{herbivory}
\alias{herbivory}
\title{Percentage of herbivore/detrivore species in the newtork}
\usage{
herbivory(graph)
}
\arguments{
\item{graph}{an \code{igraph} object.}
}
\value{
\code{integer} the percentage of herbivore/detrivore species.
}
\description{
Percentage of herbivore/detrivore species in the newtork
}
\references{
Baiser, B., Gotelli, N. J., Buckley, H. L., Miller, T. E., & Ellison, A. M. (2012). Geographic variation in network structure of a nearctic aquatic food web. Global Ecology and Biogeography.
}
