% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int.R
\name{inter}
\alias{inter}
\title{Percentage of intermediate taxa}
\usage{
inter(graph)
}
\arguments{
\item{graph}{}
}
\value{
numeric: percentage of intermediate taxa
}
\description{
Percentage of taxa with both consumer and ressources.
}
