% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy.R
\name{link}
\alias{link}
\title{Link}
\usage{
link(graph)
}
\arguments{
\item{graph}{an igraph object.}
}
\value{
dataframe: number of links of each taxon.
}
\description{
Number of consumers and resources per taxon.
}
\references{
Baiser, B., Gotelli, N. J., Buckley, H. L., Miller, T. E., & Ellison, A. M. (2012). Geographic variation in network structure of a nearctic aquatic food web. Global Ecology and Biogeography.
}
