% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{link_density}
\alias{link_density}
\title{Linkage density}
\usage{
link_density(graph)
}
\arguments{
\item{graph}{an igraph object.}
}
\value{
numeric: the linkage denisty of the graph.
}
\description{
Number of link per species.
}
\examples{
data(aleutian)

link_density(aleutian)
}
\references{
Delmas, E., Besson, M., Brice, M.-H., Burkle, L. A., Dalla Riva, G. V., Fortin, M.-J., … Poisot, T. (2019). Analysing ecological networks of species interactions. Biological Reviews.
}
