% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy.R
\name{link_mean}
\alias{link_mean}
\title{Mean number of links}
\usage{
link_mean(graph, std = FALSE)
}
\arguments{
\item{graph}{an igraph object.}

\item{std}{should the number of links per taxon be standardized by 1/(LS)? Value must be `TRUE` or `FALSE`.}
}
\value{
the mean number of consumers and resources per taxon for the whole food web.
}
\description{
Mean number of consumers and resources per taxon for the whole food web.
}
\references{
Williams, R. J., & Martinez, N. D. (2000). Simple rules yield complex food webs. Nature.
}
