% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy.R
\name{link_sd}
\alias{link_sd}
\title{Normalized standard deviation of links}
\usage{
link_sd(graph)
}
\arguments{
\item{graph}{an igraph object.}
}
\value{
dataframe: standardized number of links of each taxon.
}
\description{
Number of consumers and resources per taxon standardized by L/S.
}
\examples{
data(aleutian)

link_sd(aleutian)
}
\references{
Williams, R. J., & Martinez, N. D. (2000). Simple rules yield complex food webs. Nature.
}
