% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nb_edge.R
\name{nb_edge}
\alias{nb_edge}
\title{Number of edge of a graph.}
\usage{
nb_edge(igraph)
}
\arguments{
\item{igraph}{object.}
}
\value{
`integer` The number of edge of the graph.
}
\description{
Number of edge of a graph.
}
