% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestedness.R
\name{nestedness}
\alias{nestedness}
\alias{nestedness_total}
\title{Nestedness}
\usage{
nestedness(graph, transpose = FALSE)

nestedness_total(graph)
}
\arguments{
\item{graph}{an \code{igraph} object.}

\item{transpose}{\code{logical}. If \code{FALSE} the nestedness of rows with regard to the columns is comptute. If \code{TRUE} the nestedness of columns with regard to rows is compute.}
}
\value{
\code{numeric} the nestedness.
}
\description{
Nestedness of a graph.
}
\details{
\code{nestedness()} return the nestedness of a network rows with regard to the columns or vice-versa.

\code{nestedness_total()} return the whole nestedness as the mean of the two values.
}
