% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omnivor-package.R
\docType{package}
\name{omnivor-package}
\alias{omnivor}
\alias{omnivor-package}
\title{omnivor: Ecological Networks Analysis}
\description{
Function for the analysis of ecological networks designed for ecologists.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/clementviolet/omnivor}
  \item Report bugs at \url{https://github.com/clementviolet/omnivor/issues}
}

}
\author{
\strong{Maintainer}: Clément Violet \email{clement.violet@etudiant.univ-brest.fr} (0000-0001-6217-5891)

Authors:
\itemize{
  \item Dominique Gravel \email{dominique.gravel@usherbrooke.ca}
}

}
\keyword{internal}
