% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omnivore.R
\name{omnivore_avg_prey_tl}
\alias{omnivore_avg_prey_tl}
\title{Retrive if a taxon is omnivore}
\usage{
omnivore_avg_prey_tl(graph)
}
\arguments{
\item{graph}{an igraph object.}
}
\value{
dataframe. First column containing the taxa and second if the taxon is omnivore or not.
}
\description{
Retrive if a taxon is omnivore or not, based on average prey trophic level method.
}
\keyword{internal}
