% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omnivore.R
\name{omnivore_short_tl}
\alias{omnivore_short_tl}
\title{Retrive if a taxon is omnivore.}
\usage{
omnivore_short_tl(graph)
}
\arguments{
\item{graph}{an igraph object.}
}
\value{
dataframe. First column containing the taxa and second if the taxon is omnivore or not.
}
\description{
Retrive if a taxon is omnivore or not, based on shortest trophic level method.
}
\keyword{internal}
