% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{path_length}
\alias{path_length}
\title{Characteristic path length,}
\usage{
path_length(graph, directed = TRUE, ...)
}
\arguments{
\item{graph}{an igraph object.}

\item{directed}{logical: is the graph directed?}

\item{...}{others arguments passed to \code{igraph::mean_distance}.}
}
\value{
numeric: the characteristic path length.
}
\description{
Characteristic path length,
}
\references{
Baiser, B., Gotelli, N. J., Buckley, H. L., Miller, T. E., & Ellison, A. M. (2012). Geographic variation in network structure of a nearctic aquatic food web. Global Ecology and Biogeography.
}
