% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trophic_length.R
\name{prey_avg_tl}
\alias{prey_avg_tl}
\title{Prey-averaged trophic level}
\usage{
prey_avg_tl(graph)
}
\arguments{
\item{graph}{an igraph object.}
}
\value{
dataframe contening the taxons and the trophic level.
}
\description{
Prey-averaged trophic level is calculated as follow :
\deqn{TL_j = 1 + \sum_{i = 1}{S}l_{ij}\frac{TL_{i}}{n_j}}
With :
\itemize{
    \item \eqn{TL_j} the trophic level of the \eqn{j} specie;
    \item \eqn{S} the number of species of the trophque chain;
    \item \eqn{l_{ij}} equal 1 if the specie \eqn{j} consume specie \eqn{i};
    \item \eqn{n_j} the number of specie consumed by the specie \eqn{j}.
}
}
\references{
Williams, R.J. & Martinez, N.D., 2004. Limits to Trophic Levels and Omnivory in Complex Food Webs: Theory and Data. The American Naturalist
}
