% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top.R
\name{top}
\alias{top}
\title{Percentage of top species}
\usage{
top(graph)
}
\arguments{
\item{graph}{an igraph object.}
}
\value{
numeric: percentage of top species.
}
\description{
Percentage of species without consumers.
}
\examples{
data(aleutian)

top(aleutian)
}
\references{
Baiser, B., Gotelli, N. J., Buckley, H. L., Miller, T. E., & Ellison, A. M. (2012). Geographic variation in network structure of a nearctic aquatic food web. Global Ecology and Biogeography.
}
