% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy.R
\name{vul}
\alias{vul}
\title{Vulnerability}
\usage{
vul(graph)
}
\arguments{
\item{graph}{an igraph object.}
}
\value{
dataframe: vulnerability of each taxon.
}
\description{
The number of predators of a taxon.
}
\examples{
data(aleutian)

vul(aleutian)
}
\references{
Williams, R. J., & Martinez, N. D. (2000). Simple rules yield complex food webs. Nature.
}
