% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy.R
\name{vul_mean}
\alias{vul_mean}
\title{Mean number of vulnerability}
\usage{
vul_mean(graph, std = FALSE)
}
\arguments{
\item{graph}{an igraph object.}

\item{std}{should the number of predator per taxon be standardized by 1/(LS)? Value must be `TRUE` or `FALSE`.}
}
\description{
Mean number of predators of a taxon for the whole food web.
}
\references{
Williams, R. J., & Martinez, N. D. (2000). Simple rules yield complex food webs. Nature.
}
