% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy.R
\name{vul_sd}
\alias{vul_sd}
\title{Normalized standard deviation of vulnerability}
\usage{
vul_sd(graph)
}
\arguments{
\item{graph}{an igraph object.}
}
\value{
numeric: the normalized standard deviation of vulnerability.
}
\description{
The number of predators of a taxon standardized by L/S.
}
\references{
Baiser, B., Gotelli, N. J., Buckley, H. L., Miller, T. E., & Ellison, A. M. (2012). Geographic variation in network structure of a nearctic aquatic food web. Global Ecology and Biogeography.
}
