% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/can.R
\name{can}
\alias{can}
\title{Percentage of cannibals}
\usage{
can(graph)
}
\arguments{
\item{graph}{an igraph object.}
}
\value{
numeric: Percentage of cannibals
}
\description{
Percentage of cannibals species in a food web.
}
