% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectance.R
\name{connectance}
\alias{connectance}
\title{Connectance of a graph}
\usage{
connectance(graph, loops = TRUE, directed = TRUE)
}
\arguments{
\item{graph}{an igraph object.}

\item{loops}{logical: are species able to interact with themselves?}

\item{directed}{logical: are interactions directed?}
}
\value{
numeric: the value of the connectance.
}
\description{
Proportion of all possible links that are realized.
}
\references{
Delmas, E., Besson, M., Brice, M.-H., Burkle, L. A., Dalla Riva, G. V., Fortin, M.-J., … Poisot, T. (2019). Analysing ecological networks of species interactions. Biological Reviews.
}
