% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectance.R
\name{connectance}
\alias{connectance}
\title{Connectance of a graph}
\usage{
connectance(graph, loops = TRUE, directed = TRUE)
}
\arguments{
\item{graph}{an igraph object.}

\item{loops}{logical: are species able to interact with themselves?}

\item{directed}{logical: are interactions directed?}
}
\value{
numeric: the value of the connectance.
}
\description{
Proportion of all possible links that are realized. In a directed unipartite network \eqn{C = \frac{L}{S^2}}, in a directed newtork where species cannot interact with themselves \eqn{C = \frac{L}{S(S-1)}}. In undirected newtork, if species cannot interact with themselves \eqn{C = \frac{L}{S\frac{S-1}{2}}} otherwise \eqn{C = \frac{L}{S\frac{S-1}{2}}}. Where \eqn{L} is the number of links and \eqn{S} the number of species.
}
\examples{
data(aleutian)

connectance(aleutian)

connectance(aleutian, loops = FALSE)

connectance(aleutian, directed = FALSE)
}
\references{
Delmas, E., Besson, M., Brice, M.-H., Burkle, L. A., Dalla Riva, G. V., Fortin, M.-J., … Poisot, T. (2019). Analysing ecological networks of species interactions. Biological Reviews.
}
