% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy.R
\name{link}
\alias{link}
\title{Link}
\usage{
link(graph)
}
\arguments{
\item{graph}{an igraph object.}
}
\value{
dataframe: number of links of each taxon.
}
\description{
Number of consumers and resources per taxon.
}
\examples{
data(aleutian)

link(aleutian)
}
\references{
Williams, R. J., & Martinez, N. D. (2000). Simple rules yield complex food webs. Nature.
}
