% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{link_density}
\alias{link_density}
\title{Linkage density}
\usage{
link_density(igraph)
}
\arguments{
\item{igraph}{}
}
\value{
A numeric, the linkage denisty of the graph.
}
\description{
Linkage density
}
\references{
Delmas, E., Besson, M., Brice, M.-H., Burkle, L. A., Dalla Riva, G. V., Fortin, M.-J., … Poisot, T. (2019). Analysing ecological networks of species interactions. Biological Reviews.
}
