% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy.R
\name{link_sd}
\alias{link_sd}
\title{Normalized standard deviation of links}
\usage{
link_sd(graph)
}
\arguments{
\item{graph}{an igraph object.}
}
\description{
Number of consumers and resources per taxon standardized by L/S.
}
\references{
Baiser, B., Gotelli, N. J., Buckley, H. L., Miller, T. E., & Ellison, A. M. (2012). Geographic variation in network structure of a nearctic aquatic food web. Global Ecology and Biogeography.
}
