% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food_chain_length.R
\name{mean_tl}
\alias{mean_tl}
\title{Mean trophic level}
\usage{
mean_tl(graph, method = c("shortest", "average_prey", "weighted"))
}
\arguments{
\item{method}{computation method for the trophic level. Must be one of `shortest`, `average_prey` `weighted`. See \code{\link{short_tl}},\code{\link{prey_avg_tl}}, \code{\link{short_wght_tl}} for more informations.}
}
\value{
numeric: the mean trophic level
}
\description{
Mean trophic level for the food web. Also called "mean food chain length".
}
\references{
Baiser, B., Gotelli, N. J., Buckley, H. L., Miller, T. E., & Ellison, A. M. (2012). Geographic variation in network structure of a nearctic aquatic food web. Global Ecology and Biogeography.
}
