% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nb_edge.R
\name{nb_edge}
\alias{nb_edge}
\title{Number of edge of in a graph}
\usage{
nb_edge(graph)
}
\arguments{
\item{graph}{an igraph object.}
}
\value{
integer: the number of edge of the graph.
}
\description{
Number of edge of in a graph.
}
\references{
Delmas, E., Besson, M., Brice, M.-H., Burkle, L. A., Dalla Riva, G. V., Fortin, M.-J., … Poisot, T. (2019). Analysing ecological networks of species interactions. Biological Reviews.
}
