% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nb_node.R
\name{nb_node}
\alias{nb_node}
\title{Number of nodes of in a graph}
\usage{
nb_node(graph)
}
\arguments{
\item{graph}{an igraph object.}
}
\value{
integer: the number of node of the graph.
}
\description{
Number of nodes of a graph.
}
\references{
Delmas, E., Besson, M., Brice, M.-H., Burkle, L. A., Dalla Riva, G. V., Fortin, M.-J., … Poisot, T. (2019). Analysing ecological networks of species interactions. Biological Reviews.
}
