% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy.R
\name{vul}
\alias{vul}
\title{Vulnerability}
\usage{
vul(graph)
}
\arguments{
\item{graph}{an igraph object.}
}
\value{
dataframe: vulnerability of each taxon.
}
\description{
The number of predators of a taxon.
}
\references{
Baiser, B., Gotelli, N. J., Buckley, H. L., Miller, T. E., & Ellison, A. M. (2012). Geographic variation in network structure of a nearctic aquatic food web. Global Ecology and Biogeography.
}
