% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy.R
\name{vul_sd}
\alias{vul_sd}
\title{Normalized standard deviation of vulnerability}
\usage{
vul_sd(graph)
}
\arguments{
\item{graph}{an igraph object.}
}
\value{
numeric: the normalized standard deviation of vulnerability.
}
\description{
The number of predators of a taxon standardized by L/S.
}
\examples{
data(aleutian)

vul_sd(aleutian)
}
\references{
Williams, R. J., & Martinez, N. D. (2000). Simple rules yield complex food webs. Nature.
}
