% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omnivore.R
\name{which_omnivore}
\alias{which_omnivore}
\title{Retrive if a taxon is omnivore}
\usage{
which_omnivore(graph, method = c("shortest", "average_prey"))
}
\arguments{
\item{graph}{an igraph object.}

\item{method}{computation method for the trophic level. Must be one of `shortest` or `average_prey`. See \code{\link{short_tl}} and \code{\link{prey_avg_tl}} for more informations.}
}
\value{
dataframe. First column containing the taxa and second if the taxon is omnivore or not.
}
\description{
Retrive if a taxon is omnivore or not, based on average prey trophic level or shortest trophic level method.
}
\references{
Baiser, B., Gotelli, N. J., Buckley, H. L., Miller, T. E., & Ellison, A. M. (2012). Geographic stats::variation in network structure of a nearctic aquatic food web. Global Ecology and Biogeography.
}
